put templates in their own file and add indentation wrapper from Matthias Riegler
